24_telegraph/requirements.txt
alembic==1.0.7
awesome-slugify==1.6.5
Click==7.0
Flask==1.0.2
Flask-Migrate==2.3.1
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
psycopg2==2.7.7
python-dateutil==2.7.5
python-dotenv==0.10.1
python-editor==1.0.3
regex==2019.1.24
six==1.12.0
SQLAlchemy==1.2.17
Unidecode==0.4.21
Werkzeug==0.14.1
